refactor: 이미지 슬라이드 네비게이션 모듈화(hooks)
